UPLOAD EXCEL OK
PUEDE RECIBIR Y ANALIZAR EL EXCEL QUE LE MANDO DESDE EL FRONTEND

+++ src/main/java/com/backend/APIRest/repository/checador/HoraRepository.java
@@ -4,6 +4,10 @@ import com.backend.APIRest.model.entidades.checador.Hora;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.backend.APIRest.model.entidades.checador.Hora;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
 @Repository
-public interface HoraRepository extends JpaRepository<Hora, Long> {
+public interface HoraRepository extends JpaRepository<Hora, Integer> {
 }
